+++ pyriemann/embedding.py
@@ -7,9 +7,14 @@ from scipy.linalg import solve, eigh
 from scipy.sparse import csr_matrix
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.manifold import spectral_embedding
+from sklearn.manifold._utils import _binary_search_perplexity
+from time import time 
 
 from .utils.distance import pairwise_distance
 from .utils.kernel import kernel as kernel_fct
+from .utils.base import sqrtm, invsqrtm, logm
+from .datasets import sample_gaussian_spd
+from .utils.gradient import retraction, norm_SPD
 
 
 class SpectralEmbedding(BaseEstimator):
@@ -273,7 +278,7 @@ class LocallyLinearEmbedding(TransformerMixin, BaseEstimator):
         Parameters
         ----------
         X : ndarray, shape (n_matrices, n_channels, n_channels)
-            Set of SPD matrices.
+            
         y : None
             Not used, here for compatibility with sklearn API.
 
@@ -438,6 +443,308 @@ def locally_linear_embedding(
     return embd, error
 
 
+
+class tSNE(BaseEstimator):
+    """AIRM Riemannian version of t-SNE algorithm.
+
+    This Riemannian version of the t-SNE algorihtm was presented in [1]_.
+    Reduces a set of d x d SPD matrices into a set of 2 x 2 SPD matrices.
+
+    Parameters
+    ----------
+    n_components : int, default = 2
+        Dimension of the matrices in the embedded space.
+    perplexity : int, default = None
+        Perplexity used in the Riemannian t-SNE algorithm. 
+        If perplexity = None, it will be set to 0.75*n_matrices
+    verbosity : int, default = 0
+        Level of information printed by the optimizer while it operates: 
+        0 is silent, 2 is most verbose.
+    max_it : int, default = 10000
+        Maximum number of iterations used for the Riemannian gradient 
+        descent.
+    max_time : int default = 300
+        Maximum time on the run time of the Riemannian gradient descent 
+        in seconds.
+    random_state : int, default=None
+        Pass an int for reproducible output across multiple function calls.
+
+    Attributes
+    ----------
+    embedding_ : ndarray, shape (n_matrices, n_components, n_components)
+        The result of the optimization process.
+    
+    References
+    ----------
+    .. [1] `Geometry-Aware visualization of high dimensional Symmetric Positive Definite matrices
+        <https://openreview.net/pdf?id=DYCSRf3vby>`_
+        T. de Surrel, S. Chevallier, F. Lotte and F. Yger. 
+        Transactions on Machine Learning Research, 2025
+    """
+
+    def __init__(
+        self,
+        n_components = 2,
+        perplexity=None,
+        verbosity=0,
+        max_it=10000,
+        max_time=100,
+        random_state=None, 
+    ):
+        self.n_components = n_components
+        self.perplexity = perplexity
+        self.verbosity = verbosity
+        self.max_it = max_it
+        self.max_time = max_time
+        self.random_state = random_state
+
+    def _compute_similarties(self, X):
+        """Computed the high dimensional symmetrized conditional similarities 
+        p_{ij} for the t-SNE algorithm.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices to reduce.
+
+        Returns
+        -------
+        P : ndarray, shape (n_matrices, n_matrices)
+            The matrix of the symmetrized conditional probabilities of X.
+        """
+        nb_mat = X.shape[0]
+        Dsq = pairwise_distance(X, squared=True)
+        Dsq = Dsq.astype(np.float32, copy=False)
+        # Use _binary_search_perplexity from sklearn to compute conditional probabilities
+        # such that they approximately match the desired perplexity
+        conditional_P = _binary_search_perplexity(Dsq, self.perplexity, 0)
+
+        # Symmetrize the conditional probabilities
+        P = conditional_P + conditional_P.T
+        return P / (2 * nb_mat)
+
+    def _compute_low_affinities(self, Y):
+        """Computed the low dimensional similarities q_{ij} for the t-SNE 
+        algorithm.
+
+        Parameters
+        ----------
+        Y : ndarray, shape (n_matrices, n_components, n_components)
+            Set of SPD matrices.
+
+        Returns
+        -------
+        Q : ndarray, shape (n_matrices, n_matrices)
+            The matrix of the low dimensional similarities conditional
+            probabilities of Y.
+        Dsq : ndarray, shape (n_matrices, n_matrices)
+            The array containing the squared Riemannian distances between
+            the points in X.
+        """
+        nb_mat = Y.shape[0]
+        Dsq = pairwise_distance(Y, squared=True)
+
+        denominator = np.sum(
+            [np.sum([np.delete(1 / (1 + Dsq[k, :]), k)]) for k in range(nb_mat)]
+        )
+        Q = 1 / (1 + Dsq) / denominator
+        np.fill_diagonal(Q, 0)
+        return Q, Dsq
+
+    def _cost(self, P, Q):
+        """Computed the loss of the t-SNE, that is the Kullback-Leibler 
+        divergence between P and Q.
+
+        Parameters
+        ----------
+        P : ndarray, shape (n_matrices, n_matrices)
+            The matrix of the symmetrized conditional probabilities of X.
+        Q : ndarray, shape (n_matrices, n_matrices)
+            The matrix of the low dimensional similarities conditional 
+            probabilities of Y.
+
+        Returns
+        -------
+        _ : float
+            The cost of the t-SNE.
+        """
+        return np.sum(P * np.log((P + np.eye(P.shape[0])) / (Q + np.eye(P.shape[0]))))
+
+    def _riemannian_gradient(self, Y, P, Q, Dsq):
+        """Computed the Riemannian gradient of the loss of the t-SNE.
+
+        Parameters
+        ----------
+        Y : ndarray, shape (n_matrices, n_components, n_components)
+            Set of SPD matrices.
+        P : ndarray, shape (n_matrices, n_matrices)
+            The matrix of the symmetrized conditional probabilities of X.
+        Q : ndarray, shape (n_matrices, n_matrices)
+            The matrix of the low dimensional similarities conditional 
+            probabilities of Y.
+        Dsq : ndarray, shape (n_matrices, n_matrices)
+            The Riemannian distance matrix of Y.
+
+        Returns
+        -------
+        grad : ndarray, shape (n_matrices, n_components, n_components)
+            The Riemannian gradient of the cost of the t-SNE.
+        """
+        nb_mat = P.shape[0]
+        grad = np.zeros((nb_mat, self.n_components, self.n_components))
+        Y_i_invsqrt = invsqrtm(Y)
+        Y_i_sqrt = sqrtm(Y)
+        for i in range(nb_mat):
+            log_riemann = (
+                Y_i_sqrt[i] @ logm(Y_i_invsqrt[i] @ Y @ Y_i_invsqrt[i]) @ Y_i_sqrt[i]
+            )
+            grad[i] = -4 * np.sum(
+                ((P[i] - Q[i]) / (1 + Dsq[i]))[:, np.newaxis, np.newaxis] * log_riemann,
+                axis=0,
+            )
+        return grad
+
+    def _run_minimization(self, P):
+        """Run the minimization to solve the t-SNE optimization.
+
+        Parameters
+        ----------
+        P : ndarray, shape (n_matrices, n_matrices)
+            The matrix of the symmetrized conditional probabilities of X.
+
+        Returns
+        -------
+        current_sol : ndarray, shape (n_matrices, n_components, n_components)
+            The solution of the t-SNE problem.
+        """
+        tol_step = 1e-6
+        current_sol = self.initial_point
+        self.loss_evolution = []
+        initial_time = time()
+
+        # loop over iterations
+        for i in range(self.max_it):
+            if self.verbosity >= 2 and i % 100 == 0:
+                print("Iteration : ", i)
+
+            # get the current value for the loss function
+            Q, Dsq = self._compute_low_affinities(current_sol)
+            loss = self._cost(P, Q)
+            self.loss_evolution.append(loss)
+
+            # get the direction of steepest descent
+            direction = self._riemannian_gradient(current_sol, P, Q, Dsq)
+            norm_direction = norm_SPD(current_sol, direction)
+
+            # backtracking line search
+            if i == 0:
+                alpha = 1.0 / norm_direction
+            else:
+                # Pick initial step size based on where we were last time and look a bit further
+                # See Boumal, 2023, Section 4.3 for more insights.
+                alpha = 4 * (self.loss_evolution[-2] - loss) / (norm_direction**2)
+
+            tau = 0.50
+            r = 1e-4
+            maxiter_linesearch = 25
+
+            retracted = retraction(current_sol, -alpha * direction)
+            Q_retracted, Dsq_retracted = self._compute_low_affinities(retracted)
+            loss_retracted = self._cost(P, Q_retracted)
+
+            # Backtrack while the Armijo criterion is not satisfied
+            for _ in range(maxiter_linesearch):
+                if loss - loss_retracted > r * alpha * norm_direction**2:
+                    break
+                alpha = tau * alpha
+
+                retracted = retraction(current_sol, -alpha * direction)
+                Q_retracted, Dsq_retracted = self._compute_low_affinities(retracted)
+                loss_retracted = self._cost(P, Q_retracted)
+            else:
+                warnings.warn("Maximum iteration in linesearched reached.")
+
+            # update variable for next iteration
+            current_sol = retracted
+
+            # test if the step size is small
+            crit = norm_SPD(current_sol, -alpha * direction)
+            if crit <= tol_step:
+                if self.verbosity >= 1:
+                    print("Min stepsize reached")
+                break
+
+            # test if the maximum time has been reached
+            if time() - initial_time >= self.max_time:
+                warnings.warn("Time limite reached after " + str(i) + " iterations.")
+                break
+
+        else:
+            warnings.warn("Maximum iterations reached.")
+        if self.verbosity >= 1:
+            print(
+                "Optimization done in {:.2f} seconds.".format(time() - initial_time)
+            )
+        return current_sol
+
+    def fit(self, X, y=None):
+        """Fit X to smaller SDP matrices using the Riemannian t-SNE algorithm.
+
+        Parameters
+        ----------
+        X : array_like of shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : None
+            Not used, here for compatibility with sklearn API.
+
+        Returns
+        -------
+        res_opti :  ndarray, shape (n_matrices, n_components, n_components)
+            The solution of the t-SNE problem.
+        """
+        (n_matrices, _, _) = X.shape
+
+        if self.perplexity is None:
+            self.perplexity = int(0.75 * n_matrices)
+
+        # Compute similarities in the high dimension space
+        P = self._compute_similarties(X)
+
+    
+        # Sample initial solution close to the identity
+        sigma = 1
+        self.initial_point = sample_gaussian_spd(
+            n_matrices, 
+            mean=np.eye(self.n_components), 
+            sigma=sigma, 
+            random_state = self.random_state
+        )
+        if self.verbosity >= 1:
+            print("Optimizing...")
+        self.embedding_ = self._run_minimization(P)
+
+        return self
+    
+    def fit_transform(self, X, y=None):
+        """Calculate the coordinates of the embedded matrices 
+        using t-SNE.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : None
+            Not used, here for compatibility with sklearn API.
+
+        Returns
+        -------
+        X_new : ndarray, shape (n_matrices, n_components, n_components)
+            Coordinates of embedded matrices using t-SNE.
+
+        """
+        self.fit(X)
+        return self.embedding_
+    
 def _check_dimensions(X, Y=None, n_components=None, n_neighbors=None):
     n_matrices, n_channels, n_channels = X.shape